Port ServicingPipeline.ps1 to ProjectsV2
GraphQL was a steep learning curve, but I think I get it now.
I've written the queries to go deep into objects to find the exact merge
commit (for a PR) or the PR->merge commit chain (for an issue). This
means that ServicingPipeline now supports *issues* in project boards
rather than just pull requests!

I am sure that it is going to break in some way, eventually.