Clip the "current commandline" at the cursor position
This is particularly relevant to pwsh with the "ghost text" enabled. In that
scenario, pwsh writes out the predicted command to the right of the cursor. With
`showSuggestions(useCommandline=true)`, we'd auto-include that text in the
filter, and that was effectively useless.

This instead defaults us to not use anything to the right of the cursor
(inclusive) for what we consider "the current commandline"

closes #17772